v2.2

+++ Backend/Controllers/RecipeManagerController.cs
@@ -6,12 +6,22 @@ using Microsoft.AspNetCore.Http;
 using Microsoft.AspNetCore.Mvc;
 using Backend.Models.BindingModels;
 using Backend.Models;
+using Microsoft.AspNetCore.Authorization;
+using Microsoft.Extensions.Logging;
+using Newtonsoft.Json;
 
 namespace Backend.Controllers
 {
     [ApiController]
     public class RecipeManagerController : ControllerBase
     {
+        private readonly ILogger<RecipeManagerController> _logger;
+
+        public RecipeManagerController(ILogger<RecipeManagerController> logger)
+        {
+            this._logger = logger;
+        }
+
         /// <summary>
         /// Добавить новый рецепт
         /// </summary>
@@ -20,13 +30,18 @@ namespace Backend.Controllers
         /// <response code="200">Рецепт добавлен</response>
         /// <response code="400">Некорректные значения</response>
         /// <response code="500">Внутренняя ошибка (читать сообщение в ответе)</response>
+        /// <response code="401">Неавторизован или низкий уровень доступа</response>
         [HttpPost("api/[controller]/add")]
+        [Authorize(Roles = "admin, open")]
         public IActionResult Post(FullInfoRecipeBindingModel recipe)
         {
+            _logger.LogError("Add: запуск с параметрами\n" + JsonConvert.SerializeObject(recipe));
             try
             {
                 if (recipe == null)
                     return BadRequest();
+
+                recipe.author = User.Identity.Name;
 
                 if(AddIntoDB(recipe))
                 {
@@ -39,7 +54,7 @@ namespace Backend.Controllers
             }
             catch(Exception e)
             {
-                return StatusCode(500, e.InnerException.Message);
+                return StatusCode(500, e.Message);
             }
         }
 
@@ -69,7 +84,17 @@ namespace Backend.Controllers
                         {
                             IngredientList i = new IngredientList();
                             i.Idrecipe = curidrecipe;
-                            i.Idingredient = ingredient.idIngredientList;
+                            i.Idingredient = ingredient.idIngredient;
+                            //if(_model.Ingredient.ToList().FindAll(i => i.Name.ToLower() == ingredient.ingredient.ToLower()).Count() == 0)
+                            //{
+                            //    _model.Ingredient.Add(new Ingredient { Name = ingredient.ingredient });
+                            //    _model.SaveChanges();
+                            //    i.Idingredient = _model.Ingredient.ToList().Last().Idingredient;
+                            //}
+                            //else
+                            //{
+                            //    i.Idingredient = _model.Ingredient.ToList().Find(i => i.Name.ToLower() == ingredient.ingredient.ToLower()).Idingredient;
+                            //}
                             i.Amount = ingredient.amount;
                             _model.IngredientList.Add(i);
                         }
@@ -101,25 +126,36 @@ namespace Backend.Controllers
         /// </summary>
         /// <param name="recipe">Данные рецепта</param>
         /// <returns>Результат, получилось изменить или нет</returns>
-        /// <response code="400">Некорректные значения</response>
         /// <response code="200">Данные рецепта изменены</response>
+        /// <response code="400">Некорректные значения</response>
+        /// <response code="406">Логин пользователя и записанный автор рецепта не совпадает</response>
         /// <response code="500">Внутренняя ошибка (читать сообщение в ответе)</response>
+        /// <response code="401">Неавторизован или низкий уровень доступа</response>
         [HttpPost("api/[controller]/update")]
+        [Authorize(Roles = "admin, open, blocked")]
         public IActionResult Update(FullInfoRecipeBindingModel recipe)
         {
+            _logger.LogError("Update: запуск с параметрами\n" + JsonConvert.SerializeObject(recipe));
             try
             {
-                if (recipe == null)
+                if (recipe.idRecipe < 1)
                     return BadRequest();
 
-                if (UpdateInDB(recipe))
+                if(CheckAuthorLogin(User.Identity.Name, recipe.idRecipe) || User.IsInRole("admin"))
                 {
-                    return Ok();
+                    var response = UpdateInDB(recipe);
+
+                    if (response == "OK")
+                    {
+                        return Ok();
+                    }
+                    else
+                    {
+                        return StatusCode(500, "Во время изменения что-то пошло не так!" + response);
+                    }
                 }
-                else
-                {
-                    return StatusCode(500, "Во время изменения что-то пошло не так");
-                }
+
+                return StatusCode(406);
             }
             catch (Exception e)
             {
@@ -127,7 +163,7 @@ namespace Backend.Controllers
             }
         }
 
-        private bool UpdateInDB(FullInfoRecipeBindingModel recipe)
+        private string UpdateInDB(FullInfoRecipeBindingModel recipe)
         {
             using (ModelDbContext _model = new ModelDbContext())
             {
@@ -137,16 +173,23 @@ namespace Backend.Controllers
                     {
                         Recipe r = _model.Recipe.ToList().Find(r => r.Idrecipe == recipe.idRecipe);
 
-                        if (recipe.name != null)
+                        if (recipe.name != null && recipe.name.Length > 0)
                             r.Name = recipe.name;
+                        else
+                            throw new Exception("Имя нулевое");
 
                         if (recipe.idIngredient != 0)
                             r.Idingredient = recipe.idIngredient;
+                        else
+                            throw new Exception("Id главного ингредиента присвоен нулю");
+
 
                         if (recipe.idNationality != 0)
                             r.Idnationality = recipe.idNationality;
+                        else
+                            throw new Exception("Id национальности присвоено нулю");
 
-                        if (recipe.duration != null)
+                        if (recipe.duration != null || recipe.duration.Length > 0)
                             r.Duration = TimeSpan.Parse(recipe.duration);
 
                         if (recipe.steps.Count != 0)
@@ -154,39 +197,54 @@ namespace Backend.Controllers
                             foreach (var step in recipe.steps)
                             {
                                 Step s = _model.Step.ToList().Find(s => s.Idrecipe == recipe.idRecipe && s.OrderIndex == step.orderIndex);
-                                if(step.description!=null)
+                                if(step.description != null)
                                 {
                                     s.Description = step.description;
                                 }
                             }
                         }
+                        else
+                            throw new Exception("Отсутствуют шаги");
 
                         if (recipe.ingredientList.Count != 0)
                         {
+                            var old_ingredients = _model.IngredientList.ToList().FindAll(i => i.Idrecipe == recipe.idRecipe);
+
+                            _model.IngredientList.RemoveRange(old_ingredients);
+
                             foreach (var ingredient in recipe.ingredientList)
                             {
-                                IngredientList i = _model.IngredientList.ToList().Find(i => i.Idrecipe == recipe.idRecipe && i.IdingredientList == ingredient.idIngredientList);
+                                IngredientList new_ingredient = new IngredientList();
+                                new_ingredient.Amount = ingredient.amount;
+                                new_ingredient.Idingredient = ingredient.idIngredient;
+                                new_ingredient.Idrecipe = recipe.idRecipe;
 
-                                if (ingredient.idIngredientList != 0)
-                                {
-                                    i.Idingredient = ingredient.idIngredientList;
-                                }
+                                _model.IngredientList.Add(new_ingredient);
 
-                                if (ingredient.amount != null)
-                                {
-                                    i.Amount = ingredient.amount;
-                                }
+                                //IngredientList i = _model.IngredientList.ToList().Find(i => i.Idrecipe == recipe.idRecipe && i.IdingredientList == ingredient.idIngredientList);
+
+                                //if (ingredient.idIngredient != 0)
+                                //{
+                                //    i.Idingredient = ingredient.idIngredient;
+                                //}
+
+                                //if (ingredient.amount != null)
+                                //{
+                                //    i.Amount = ingredient.amount;
+                                //}
                             }
                         }
+                        else
+                            throw new Exception("Отсутствует список ингредиентов");
 
                         _model.SaveChanges();
                         transaction.Commit();
-                        return true;
+                        return "OK";
                     }
                     catch (Exception e)
                     {
                         transaction.Rollback();
-                        return false;
+                        return e.Message;
                     }
                 }
             }
@@ -197,25 +255,35 @@ namespace Backend.Controllers
         /// </summary>
         /// <param name="idRecipe">id Рецепта</param>
         /// <returns>Результат, удалён рецепт или нет</returns>
-        /// <response code="400">Некорректные значения</response>
         /// <response code="200">Рецепт удалён</response>
+        /// <response code="400">Некорректные значения</response>
+        /// <response code="406">Логин пользователя и записанный автор рецепта не совпадает</response>
         /// <response code="500">Внутренняя ошибка (читать сообщение в ответе)</response>
+        /// <response code="401">Неавторизован или низкий уровень доступа</response>
         [HttpDelete("api/[controller]/delete")]
+        [Authorize(Roles = "admin, open, blocked")]
         public IActionResult Delete(int idRecipe)
         {
+            _logger.LogError("Delete: запуск с параметрами\n" + JsonConvert.SerializeObject(idRecipe));
             try
             {
                 if (idRecipe < 1)
                     return BadRequest();
 
-                if (DeleteFromDB(idRecipe))
+                if (CheckAuthorLogin(User.Identity.Name, idRecipe) || User.IsInRole("admin"))
                 {
-                    return Ok();
+                    if (DeleteFromDB(idRecipe))
+                    {
+                        return Ok();
+                    }
+                    else
+                    {
+                        return StatusCode(500, "Во время удаления что-то пошло не так");
+                    }
                 }
-                else
-                {
-                    return StatusCode(500, "Во время удаления что-то пошло не так");
-                }
+
+                return StatusCode(406);
+                
             }
             catch(Exception e)
             {
@@ -250,6 +318,14 @@ namespace Backend.Controllers
                         return false;
                     }
                 }
+            }
+        }
+
+        private bool CheckAuthorLogin(string login, int idRecipe)
+        {
+            using (ModelDbContext _model = new ModelDbContext())
+            {
+                return login == _model.Recipe.Find(idRecipe).Author;
             }
         }
     }
